Fixes category loading
